cleaning the code up a bit because I didnt know what I was doing back then, made stuff const refs where possible, removed using namespace std, and some other small changes

+++ DigitNeuralNetwork/GaussianRNG.h
@@ -18,6 +18,6 @@ public:
 
 	float random() 
 	{
-		return dist(e2);
+		return static_cast<float>(dist(e2));
 	}
 };
